refactor(test): make binary verifiers accessible from ext plugins
