+++ core/src/util/testing.ts
@@ -23,7 +23,7 @@ import type { EventName, Events } from "../events/events.js"
 import { EventBus } from "../events/events.js"
 import { dedent, naturalList } from "./string.js"
 import pathIsInside from "path-is-inside"
-import { join, resolve } from "path"
+import { basename, join, resolve } from "path"
 import { DEFAULT_BUILD_TIMEOUT_SEC, GARDEN_CORE_ROOT, GardenApiVersion } from "../constants.js"
 import type { GitScanMode } from "../constants.js"
 import { getRootLogger } from "../logger/logger.js"
@@ -51,6 +51,14 @@ import { GlobalConfigStore } from "../config-store/global.js"
 import { isPromise } from "./objects.js"
 import { styles } from "../logger/styles.js"
 import type { ConfigTemplateConfig } from "../config/config-template.js"
+import type { PluginToolSpec, ToolBuildSpec } from "../plugin/tools.js"
+import { parse } from "url"
+import { createReadStream, createWriteStream } from "fs"
+import got from "got"
+import { createHash } from "node:crypto"
+import { pipeline } from "node:stream/promises"
+import tmp from "tmp-promise"
+import { realpath } from "fs/promises"
 
 export class TestError extends GardenError {
   type = "_test"
@@ -568,4 +576,64 @@ export function equalWithPrecision(a: number, b: number, precision: number): boo
   const diff = round(rawDiff, precision)
   const eps = Math.pow(10, -precision)
   return diff <= eps
+}
+
+export async function downloadAndVerifyHash(
+  { architecture, platform, sha256, url }: ToolBuildSpec,
+  downloadDir: string
+) {
+  // eslint-disable-next-line no-console
+  console.log(`Downloading ${platform}-${architecture} from ${url}`)
+  const parsed = parse(url)
+  const protocol = parsed.protocol
+
+  const response =
+    protocol === "file:"
+      ? createReadStream(parsed.path!)
+      : got.stream({
+          method: "GET",
+          url,
+        })
+  const downloadedHash = response.pipe(createHash("sha256"))
+
+  const artifactName = basename(url)
+  const targetExecutable = join(downloadDir, artifactName)
+  const writeStream = createWriteStream(targetExecutable)
+  await pipeline(response, writeStream)
+  // eslint-disable-next-line no-console
+  console.log(`Download completed`)
+
+  // eslint-disable-next-line no-console
+  console.log(`Verifying hash for ${artifactName}`)
+  const downloadedSha256 = downloadedHash.digest("hex")
+
+  // eslint-disable-next-line no-console
+  console.log(`Downloaded hash: ${downloadedSha256}`)
+  // eslint-disable-next-line no-console
+  console.log(`Spec hash: ${sha256}`)
+
+  expect(downloadedSha256).to.eql(sha256)
+}
+
+export const downloadBinariesAndVerifyHashes = (toolSpecs: PluginToolSpec[]) => {
+  let tmpDir: tmp.DirectoryResult
+
+  beforeEach(async () => {
+    const dir = await tmp.dir({ unsafeCleanup: true })
+    // Fully resolve path so that we don't get path mismatches in tests
+    dir.path = await realpath(dir.path)
+    tmpDir = dir
+  })
+
+  afterEach(async () => {
+    await tmpDir.cleanup()
+  })
+
+  for (const toolSpec of toolSpecs) {
+    for (const build of toolSpec.builds) {
+      it(`${toolSpec.name} ${toolSpec.version} ${build.platform}-${build.architecture}`, async () => {
+        await downloadAndVerifyHash(build, tmpDir.path)
+      })
+    }
+  }
 }
