+++ sdk/src/testing.ts
@@ -12,7 +12,7 @@ import { uuidv4 } from "@garden-io/core/build/src/util/random.js"
 import { LogLevel, RootLogger } from "@garden-io/core/build/src/logger/logger.js"
 
 export { TestGarden, getLogMessages, getRootLogMessages } from "@garden-io/core/build/src/util/testing.js"
-export { expectError } from "@garden-io/core/build/src/util/testing.js"
+export { downloadAndVerifyHash, expectError } from "@garden-io/core/build/src/util/testing.js"
 export { makeTempDir } from "@garden-io/core/build/src/util/fs.js"
 
 export const makeTestGarden = async (projectRoot: string, opts: TestGardenOpts = {}): Promise<TestGarden> => {
